refactor: RequestMapping Class 생성
