+++ src/main/java/http/controller/impl/ErrorController.java
@@ -1,6 +1,7 @@
-package http.controller;
+package http.controller.impl;
 
 import http.HttpResponse;
+import http.controller.AbstractController;
 import http.request.HttpRequest;
 
 public class ErrorController extends AbstractController {
