+++ src/main/java/webserver/RequestHandler.java
@@ -1,12 +1,9 @@
 package webserver;
 
 import http.HttpResponse;
-import http.controller.Controller;
-import http.controller.CreateUserController;
-import http.controller.ResourcesController;
+import http.controller.RequestMapping;
 import http.request.HttpRequest;
 import http.request.HttpRequestFactory;
-import org.apache.commons.collections4.map.HashedMap;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -15,16 +12,9 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.net.Socket;
-import java.util.Map;
 
 public class RequestHandler implements Runnable {
     private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
-
-    private static final Map<String, Controller> controllers = new HashedMap<>();
-
-    static {
-        controllers.put("/user/create", new CreateUserController());
-    }
 
     private Socket connection;
 
@@ -42,8 +32,8 @@ public class RequestHandler implements Runnable {
             HttpResponse response = new HttpResponse();
             DataOutputStream dos = new DataOutputStream(out);
 
-            Controller controller = controllers.getOrDefault(request.getPath(), new ResourcesController());
-            controller.service(request, response);
+            RequestMapping.getController(request.getPath())
+                    .service(request, response);
             dos.write(response.convert().getBytes());
             dos.flush();
             logger.debug("Http Response\n{}", response);
